ai-app-builder/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
redis==5.0.1
pytest==7.4.3
pathlib